chore(package.json) Add auth:generate script
chore(env) Set default NODE_ENV to development

This commit adds a default value for NODE_ENV in the env schema to
'development' if not explicitly set. This ensures that the application
defaults to the development environment if the environment variable is
missing.

Additionally, it adds a new script `auth:generate` to the
`package.json` file. This script uses `@better-auth/cli` to generate
authentication schemas based on the configuration file located at
`./src/lib/auth/auth.ts` and outputs the generated schema to
`./src/drizzle/schemas/auth/new-auth-schema.ts`. The `--yes` flag
automatically confirms any prompts during the generation process.