feat(eslintConfig): add eslint-plugin-check-file configuration
This commit introduces eslint-plugin-check-file to enforce naming
conventions for files and folders within the src directory. It adds
the plugin to the eslint configuration and defines rules for kebab-case
naming, ignoring middle extensions and tests folders.

refactor(useUser): rename useUser hook to use-user

Renames the useUser hook file to use-user.ts for consistency with
kebab-case naming conventions and to align with eslint-plugin-check-file
rules.

fix(Profile): update import path for useUser hook

Updates the import path for the useUser hook to reflect the renaming
of the file to use-user.ts.

fix(UserGreeting): update import path for useUser hook

Updates the import path for the useUser hook to reflect the renaming
of the file to use-user.ts.